Update CMakeLists.txt in RP2350 ports to fix FREERTOS_KERNEL_PATH (#20)
The CMakeLists.txt files in the RP2350 ports attempt to derive a
value for FREERTOS_KERNEL_PATH.

The existing code is based on the RP2040 port which is one level
higher in the source tree directory hierarchy than the RP2350 ports.

This commit changes the derivation of FREERTOS_KERNEL_PATH to
account for the additional level in the directory hierarchy.